Initial commit to convert to a single conn
This simplifies a lot of the logic and un-duplicates a lot of
work/abstractions that need to be done for two conns. The down-side is
that now since state management is stored alongside the data,
referencing the same node in multiple places gets kind of wonky, so
I'll need to figure out how I want to handle that.